// src/app/juegos/interfaces/juegos.interface.ts
// Generated by https://quicktype.io

export interface JuegosResponse {
  juegos: Juego[];
}

export interface Juego {
  id_juego: number;
  nombre: string;
  desarrollador: string;
  descripcion: string;
  edad_minima: number;
  cantidad_jugadores_minima: number;
  cantidad_jugadores_maxima: number;
  duracion: number;
  imagenes: string[];
}